% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_rotation.R
\name{add_noise_columns}
\alias{add_noise_columns}
\title{Add noise dimensions}
\usage{
add_noise_columns(df, n)
}
\arguments{
\item{df}{data frame}

\item{n}{number of noise dimensions to add}
}
\description{
Add n columns of white noise to data frame. Useful to compute the SNR of a classifier.
}
\examples{
add_noise_columns(df, 8)
}
\keyword{noise}
