% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_rotation.R
\name{compute_mode}
\alias{compute_mode}
\title{Mode}
\usage{
compute_mode(v)
}
\arguments{
\item{v}{input vector}
}
\description{
Compute the most frequently occurring element (mode) of a column or vector
}
\examples{
compute_mode(c(1,2,3,3,3,3,4))
compute_mode(c("this", "and", "that", "or", "this"))
}
\keyword{statistics}
