% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_rotation.R
\name{compute_tuning_parameter}
\alias{compute_tuning_parameter}
\title{Optimize tuning parameter}
\usage{
compute_tuning_parameter(
  v_complex,
  v_oob,
  fn = weights_cut,
  min_step = 1,
  step_size = 1
)
}
\arguments{
\item{v_complex}{vector of rotation complexities}

\item{v_oob}{vector of rotation OOB errors}

\item{fn}{weight function to use, currently in weights_{cut, exp, rre}}

\item{min_step}{minimal value of h to test (maximum is always R, the number of rotations)}

\item{step_size}{determines the granularity of the grid}
}
\description{
First sort rotations by complexity (primary criterium) and then perform a grid search over h
to find the value that produces the rotation weighting with lowest OOB error.
}
\examples{
compute_tuning_parameter(v_complex=c(10,4,5,4,6), v_oob=c(0.05, 0.03, 0.1, 0.06, 0.02), fn=weights_exp, 0.1, 0.1)
}
