% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_rotation.R
\name{create_random_rotation_matrix_array}
\alias{create_random_rotation_matrix_array}
\title{Create an array of random rotations (plus identity)}
\usage{
create_random_rotation_matrix_array(numRR, dimRR)
}
\arguments{
\item{numRR}{number of random rotations to create (in addition to the identity rotation)}

\item{dimRR}{dimension of the generated random rotations}
}
\description{
Create an array containing the identity rotation plus numRR random rotations of (square) dimension dimRR
}
\examples{
create_random_rotation_matrix_array(10,3)
}
\keyword{random}
\keyword{rotation}
