% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_rotation.R
\name{determine_relevant_columns}
\alias{determine_relevant_columns}
\title{Get X and Y columns}
\usage{
determine_relevant_columns(df, response, exclude = NULL, include = NULL)
}
\arguments{
\item{df}{data frame}

\item{response}{index (integer) or column name of class (response)}

\item{exclude}{vector or list of indices (integer) and/or column names that should be excluded (default: NULL)}

\item{include}{vector or list of indices (integer) and/or column names that should ONLY be included (default: NULL)}
}
\description{
Returns columns for X matrix and Y vector (and ignored cols) from the input, given set of in-/exclusions
}
\examples{
determine_relevant_columns(df, response="myY")
determine_relevant_columns(df, response="myY", exclude=c("col1", "col9"))
determine_relevant_columns(df, response=5, include=list(2,3,4, "col10"))
}
\keyword{helper}
