% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_rotation.R
\name{df_apply_random_rotations}
\alias{df_apply_random_rotations}
\title{Apply an array of random rotations to an input data frame}
\usage{
df_apply_random_rotations(df, mArray)
}
\arguments{
\item{df}{data frame}

\item{mArray}{array of rotations}
}
\description{
Create an array containing a rotated version of the input data frame for each rotation in the matrix.
The dimension of the rotation matrices in mArray must match the number of numeric columns in df.
}
\examples{
mArray <- create_random_rotation_matrix_array(5, dimRR=3)
df_apply_random_rotations(df, mArray)
}
\keyword{random}
\keyword{rotation}
