% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_rotation.R
\name{df_convert_cols_to_factor}
\alias{df_convert_cols_to_factor}
\title{Convert cols to factor}
\usage{
df_convert_cols_to_factor(df, c_conv)
}
\arguments{
\item{df}{data frame}

\item{c_conv}{column indices of cols to be converted to factor}
}
\description{
Convert specific columns to factor
}
\examples{
df_convert_cols_to_factor(df, c(1,4,9))
}
\keyword{pre-processing}
