% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_rotation.R
\name{df_convert_cols_to_numeric}
\alias{df_convert_cols_to_numeric}
\title{Convert cols to numeric}
\usage{
df_convert_cols_to_numeric(df, c_conv)
}
\arguments{
\item{df}{data frame}

\item{c_conv}{column indices of cols to be converted to numeric}
}
\description{
Convert specific columns to numeric, coercing non-numerics to NA
}
\examples{
df_convert_cols_to_numeric(df, c(1,4,9))
}
\keyword{pre-processing}
