% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_rotation.R
\name{df_exclude_cols}
\alias{df_exclude_cols}
\title{Remove cols}
\usage{
df_exclude_cols(df, c_excl)
}
\arguments{
\item{df}{data frame}

\item{c_excl}{column indices of cols to be removed from data frame}
}
\description{
Exclude specific columns from data frame
}
\examples{
df_exclude_cols(df, c(1,4))
}
\keyword{pre-processing}
