% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_rotation.R
\name{df_numeric_strings}
\alias{df_numeric_strings}
\title{Auto-detect fake string columns that should be numeric}
\usage{
df_numeric_strings(df, tn = 0.6)
}
\arguments{
\item{df}{data frame}

\item{tn}{threshold for percentage of values that can successfully converted to numeric}
}
\description{
Returns string columns that include mostly numbers with few exceptions (e.g. "N.A.")
}
\examples{
df_numeric_strings(df)
}
\keyword{pre-processing}
