% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_rotation.R
\name{df_uniqueness_properties}
\alias{df_uniqueness_properties}
\title{Auto-detect suspicious columns}
\usage{
df_uniqueness_properties(df, th = 0.9, tl = 0.85)
}
\arguments{
\item{df}{data frame}

\item{th}{threshold for percentage of unique values}

\item{tl}{threshold for percentage of most frequent identical value}
}
\description{
Detect quasi-unique string (character) and factor columns with a threshold tl,
as well as columns that include too much repetition of one value with a threshold th
}
\examples{
df_uniqueness_properties(df)
}
\keyword{pre-processing}
