% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_rotation.R
\name{generate_testing_row_indices}
\alias{generate_testing_row_indices}
\title{Split into train and test 2/2: testing (out-of-sample) indices}
\usage{
generate_testing_row_indices(n, training_rows)
}
\arguments{
\item{n}{numer of data points in total (training + testing)}

\item{ptrain}{percentage of rows in the training set (fraction)}
}
\description{
Create a random split of indices with proportion ptrain in the training set. Also see: generate_training_row_indices(n, training_rows)
}
\examples{
train <- generate_training_row_indices(nrow(df), 0.6)
test <- generate_testing_row_indices(nrow(df), train)
}
\keyword{splitting}
