% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_rotation.R
\name{impute}
\alias{impute}
\title{Impute NA values}
\usage{
impute(v, x = 0.5)
}
\arguments{
\item{v}{numeric input vector}

\item{x}{numeric replacement value (default 0.5)}
}
\description{
Replace NA values in a vector v with a given value x (default 0.5)
}
\examples{
impute(c(0.2, 0.3, NA, 0.7, 0.9))
}
\keyword{imputation}
