% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_rotation.R
\name{is_sorted}
\alias{is_sorted}
\title{Check if sorted}
\usage{
is_sorted(v)
}
\arguments{
\item{v}{input vector}
}
\description{
Determine if an input vector v is sorted (boolean)
}
\examples{
is_sorted(c(1,2,4,3))
}
\keyword{sorting}
