% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_rotation.R
\name{match_columns}
\alias{match_columns}
\title{Obtain integer indices}
\usage{
match_columns(df, vc)
}
\arguments{
\item{df}{data frame}

\item{vc}{vector or list of indices (integer) and/or column names}
}
\description{
Matches list/vector mix of names and/or indices vc to columns in a data frame
}
\examples{
match_columns(df, list("col1", 4, "col9"))
}
\keyword{indexing}
