% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_rotation.R
\name{numeric_cols}
\alias{numeric_cols}
\title{Get numeric cols}
\usage{
numeric_cols(df)
}
\arguments{
\item{df}{data frame}
}
\description{
Returns list of columns that are of numeric type
}
\examples{
numeric_cols(df)
}
\keyword{pre-processing}
