% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_rotation.R
\name{random_rotation_matrix}
\alias{random_rotation_matrix}
\title{Random rotation (without flip)}
\usage{
random_rotation_matrix(n)
}
\arguments{
\item{n}{desired dimension of rotation}
}
\description{
Generate random member of special orthogonal group SO(n)
}
\examples{
random_rotation_matrix(3)
}
\keyword{random}
\keyword{rotation}
