% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_rotation.R
\name{random_rotation_matrix_incl_flip}
\alias{random_rotation_matrix_incl_flip}
\title{Random rotation (including potential flip)}
\usage{
random_rotation_matrix_incl_flip(n)
}
\arguments{
\item{n}{desired dimension of rotation}
}
\description{
Generate random member of orthogonal group O(n)
}
\examples{
random_rotation_matrix_incl_flip(3)
}
\keyword{random}
\keyword{rotation}
