% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_rotation.R
\name{tree_height}
\alias{tree_height}
\title{Determine height of RF decision tree}
\usage{
tree_height(mytree, r = 1, d = 1)
}
\arguments{
\item{mytree}{tree representation from randomForest output rf$forest$treemap}

\item{r}{internal recursion parameter (not used in user call)}

\item{d}{internal recursion parameter (not used in user call), represents depth}
}
\description{
Recursively compute tree height from the randomForest R package tree representation
}
\examples{
rf <- randomForest(Species ~ ., iris)
tree_height(rf$forest$treemap[,,1])
}
\keyword{random}
\keyword{rotation}
