% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_rotation.R
\name{weights_cut}
\alias{weights_cut}
\title{Equal-weight top-h rotations}
\usage{
weights_cut(R, h)
}
\arguments{
\item{R}{total number of available rotations}

\item{h}{half-life tuning parameter (cutoff)}
}
\description{
Weight of rotation based on cutoff. Name in paper: cut
}
\examples{
weights_cut(20,5)
}
\keyword{random}
\keyword{rotation}
