% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_rotation.R
\name{weights_exp}
\alias{weights_exp}
\title{Exponential decay in weight for rotations}
\usage{
weights_exp(R, h)
}
\arguments{
\item{R}{total number of available rotations}

\item{h}{half-life tuning parameter}
}
\description{
Weight of rotation based on an exponential decay. Name in paper: exp
}
\examples{
weights_exp(20,5)
}
\keyword{random}
\keyword{rotation}
