% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_rotation.R
\name{weights_rre}
\alias{weights_rre}
\title{Equal-weight all rotations}
\usage{
weights_rre(R, h)
}
\arguments{
\item{R}{total number of available rotations}

\item{h}{half-life tuning parameter (ignored for equal-weight)}
}
\description{
This is just a special case of weights_cut with h=R. Name in paper: rre. This matches the weighting of the earlier paper.
}
\examples{
weights_rre(20,0)
}
\keyword{random}
\keyword{rotation}
