% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_rotation.R
\name{zero_one_rank_map}
\alias{zero_one_rank_map}
\title{Create map}
\usage{
zero_one_rank_map(v)
}
\arguments{
\item{v}{numeric input vector}
}
\description{
Create map from input vector v in [-Inf..Inf] to a rank in [0..1]
}
\examples{
zero_one_rank_map(c(-9, 0, 4, 1000))
}
\keyword{mapping}
